InAppBrowser Compatibility
The plugin does not work correctly when using the InAppBrowser. If an InAppBrowser windows is open, it tries to present itself to the Main window, which is not the top most window.
This fix searches for the topmost window and presents the PassBook plugin to that, so it appears on top of the InAppBrowser.